fix levelup ding/sound not played during server play
